[idea] fixed broken xtend tests, unified XtextCompilerComponent and XtextAutoBuilder into XtextAutoBuilderComponent

+++ intellij/org.eclipse.xtext.idea/src/org/eclipse/xtext/idea/resource/IdeaResourceSetProvider.xtend
@@ -74,7 +74,8 @@ class IdeaResourceSetProvider {
 			}
 			for (uri : localDeleted) {
 				val file = getVirtualFile(uri)
-				file.delete(requestor)
+				if (file != null && file.exists)
+					file.delete(requestor)
 			}
 		}
